makeindexk: Avoid gcc warning on 64-bit systems
git-svn-id: svn://tug.org/texlive/trunk/Build/source@29973 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/makeindexk/sortid.c
@@ -56,7 +56,7 @@ sort_idx(void)
 #endif
     idx_dc = 0;
     idx_gc = 0L;
-    qqsort(idx_key, idx_gt, sizeof(FIELD_PTR), compare);
+    qqsort(idx_key, (size_t)idx_gt, sizeof(FIELD_PTR), compare);
 #ifdef HAVE_SETLOCALE
     setlocale(LC_COLLATE, prev_locale);
 #endif
